requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Mail==0.10.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
MarkupSafe==2.1.5
openai==1.37.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.8.0
pymongo==4.8.0
python-dotenv==1.0.1
pytubefix==6.7.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
youtube-transcript-api==0.6.2
